Load Windows from Lua in parts
In this commit, the loading of Window data from Lua has changed from a
single one-fits-all loading formula to a more attribute-value like
load-one-variable-at-a-time.

I removed the UIWindowBase class, which helped pre-loading all the
window data, and I added functions for loading a colour, a box, and a
generic attribute. The attribute loading still works for loading
window-specific attributes (such as the fps for the monitor window), but
it is now usable for basically every toggle a Window can have.

In Lua, I replaced UIWindow:collectArgs() with functions to only load
the elementary objects (e.g. colours and so). I also adjusted the code
to load mostly the same stuff that was there before.

The main difference is that I cannot adjust the font size because the
tiny-game-engine's TextBox does not contain any function for that. I
will try to improve this.