Adjust font size and aspect ratio from Lua
In this commit I added the font size / aspect ratio modification that I
just also added to tiny-game-engine.

Given that this change to tiny-game-engine still needs to be committed
and also approved for inclusion through a pull request (similar to
several others), there is currently not any official state of
tiny-game-engine (either in Strata's fork or in the upstream repository)
that would compile properly, since one needs three different pull
requests that are currently pending or in preparation.