Create LuaManager, start building Console
While starting the Console I decided it was quite inconvenient that the
Lua state existed in the Game class, which is not really intended as an
active part of Strata (it should merely rotate through all Manager
classes). The only logical solution was to refactor it into a dedicated
LuaManager class that handles Lua commands from the rest of the world.

For now, it does not actually encapsulate all Lua commands: all managers
still register their own set of Lua functions. I am not certain whether
that should remain as it is, but it definitely does make life easier
that a class registers its own functions with a borrowed Lua State - as
opposed to having to register the functions one by one through the
LuaManager by passing along functions to it.

The work done initially was on the Console, but it is only half finished
right now. It may or may not work, but anyhow it is not currently being
used actively (nor is it constructed by the UIManager). With Lua being
properly organized, the next steps should be adding a reference to the
LuaManager to the UIManager, and then creating the actual Console window
to allow executing entered commands in Lua.