Prepare execution of arbitrary Lua code chunks
In this commit I added a few simple functions that will help in
executing Lua code through the console. I am not yet done (in particular
I still need to add functionality for creating the Console window, and
calling that from the Lua startup scripts) but the general framework is
somewhat more clear now.

Once the Console window is created, it will now call executeLua() on the
LuaManager (the latter was added in the preceding commit). If I
understand the Selene documentation correctly, the current method should
suffice.