Added Console window and improved SDL input
In this commit I added the Console window to the Lua startup script, and
it is now editable. It does not display entered text yet (a 'feature' to
be added in a future commit) but it can actually crash Strata, since it
will let one re-run the startup functions (e.g. load_ui()) which then
cause the font texture to be re-allocated.

The opportunity to crash the program by calling one of the few existing
Lua functions may not seem like a big step forward, but anyhow it does
prove that Selene's Lua state properly receives the pushed input and has
the ability to call Strata's defined functions. For now that is what I
wanted to see, as it is what is needed to manipulate the terrain in the
future.

While doing this, I realised that SDL doesn't actually implement any of
the 'shift' keys (i.e. any of the characters that require a Shift to be
pressed). This seems to be a built-in feature of SDL 1.2: it cannot
really deal with varying keyboard layouts properly.

That means that everyone using Strata should now go and order a keyboard
of whatever type I currently have - errr, I mean, it means that we're
not going to fix this as long as the tiny-game-engine uses SDL 1.2.

There may be advanced methods of actually detecting the keyboard layout
that I'm unaware of, but such delicate stuff is way, way ahead of
whatever I need for Strata right now. I'll be glad if Strata works for
some keyboard layouts, for a start, and as I said I don't think this
issue is going to be addressed any sooner than tiny-game-engine moves to
SDL 2.