Add text to Console, split TextBox from Window
This commit contains the rather simple step of also displaying entered
Console commands in the Console window.

While doing that, I realised that one might want to actually display the
live command at the very bottom, and then display recent commands in
reverse order in a larger Console window that contains historical
commands.

For this, one needs a textbox that works different than currently
existing textboxes, and what's more, one needs in principle multiple
textboxes. This is not something the Window was able to handle, and this
commit mostly contains steps in that direction.

The aim is that TextBox elements can be constructed separately from
Windows, and that Windows can contain more than a single TextBox object.
For this, the core TextBox functionality was split from Window into a
new ui::TextBox class (not to be confused with the tiny::draw::TextBox).
Basically, the ui::TextBox now encapsulates the tiny::draw::TextBox
object to a large extent.

At the moment, the Window is still indirectly derived from
tiny::draw::TextBox. I have not yet decided whether that is desirable
long-term, but either way the dependency is not very deep so if the
relationship should evolve from 'is-a' to 'has-a' TextBox, no real
structural changes should be required.