Start work on Console functionality
I've decided that the Console is also going to be the class that serves
as an in-game log of all kind of events (initially debug warnings and
so, but eventually also to display game messages). That may be more
suited to Chathran than to Strata proper, but it needs to be done
anyways so it might as well be done now.

I decided not to make a separate logging class since the functionality
for entering a message and letting Lua execute it can well be coupled
with functionality to display old messages. The main argument in favor
is that they will typically share the same location on screen (for
example like an MMO's chat box) and that therefore it is convenient to
contain both functionalities (entering commands and displaying log
messages) in the same object, rather than in two distinct window-like
objects.

There is not much additional functionality yet - I still have to add a
second textbox to the Console to actually perform logging. However, I
have already added a function to the UIInterface that allows one to send
messages to the Console.

For this purpose, I also added a new UIMessage class which improves the
existing UIInformation by also specifying colour, and by allowing any
number of text fragments (not just two). I think it could become a
little more intricated once stuff like font size and symbols are added.
I intend to remove the UIInformation shortly, since pairwise storage can
in any case be done just as well through the new format.

I have considered the alternatives for a moment, but in the end I
decided to put a direct reference to the Console in the UIInterface
after all. There is something to be said for not including the logger (a
low level program capability) in the UI (a purely graphical object), but
I did not have another suitable place (the ApplManager was considered,
but rejected since it should mostly concern SDL and audio and so - i.e.
external functionality), so I put it there. In any case, it is quite
logical that anyone wanting to display UI elements such as windows (i.e.
all advanced, high-level parts of Strata) are also the ones that would
be interested in graphical logging functionality. Low level functions
such as the Mesh classes of the Terrain are not really in a position to
access whatever interface, so they will continue to log messages to the
terminal from which Strata is being run.