Start on processing mouse input
While not directly required for the functioning console, I wanted to
make a start on processing mouse input. This is done along the lines of
the existing key event logic.

Mouse events are relatively simple, and consist of a position and a
button state. They occur on the topmost window (using the same ordering
for priority as the key events) that is under the mouse position, if the
mouse buttons were up. Conversely, if the mouse buttons remain down,
they should occur on the window that was at the top of the place where
the mouse buttons were pressed down. This logic is not yet in place.