Add type specifier to Lua-loaded backgrounds
The existing way of loading backgrounds from Lua merely used a single
specifier, namely for the window (or more generally the UI object) it
was to be used for. I now extended that with a 'type' specifier that can
be used to determine what exactly the texture should do, within the
scope of the object that it belongs to. This permits objects to manage
more than a single texture.

Accordingly, the Window's setBackground can now set 'backgrounds' for
more than just itself: it can also use it to make multilayered
backgrounds or textured buttons, or anything else, and the way in which
Lua scripts manage what a texture does within a window is determined by
the 'type'. The special type 'background' sets the Window's background,
but passing the name of a button should result in setting that button's
background.

Not yet implemented, but possible, is passing the texture to a child
class of the Window. This would allow setting an arbitrary texture of a
Window-derived class through a generic interface.