Rendered Buttons
In this commit I implemented rendering of Buttons, as textboxes that
have backgrounds and that display text.

In achieving this, I had to adjust everything from Lua through the UI to
the Window and the TextBox. Many things were initially done only for
Windows, and they are now extended to apply not only to Windows
themselves but also to any TextBoxes the Window may control. This also
includes updating the Window, where the Buttons are now also updated and
have their Renderable managed through the Window.

One new addition is the ability to set Button attributes. Currently this
functionality, through the UI, only sets buttons that belong to Windows
but in principle any button could have its attributes adjusted like
this. If there is ever a non-Window UI object that has buttons, its
management will be analogous to that of Windows now.

On the Lua side, the main addition is the UIButton object that defines a
button.

With the button now visible, it still doesn't really respond to clicks,
and its text isn't as flexible (or multicolored) as it was. To solve the
latter, I want to add color encoding such that in Lua itself colors can
be set through the text fragment itself (e.g. via #a844ff; encoding),
such that one doesn't need to set various text fragments and call
functions in between every two fragments when a color change is due.

The click response is also still work in progress, but now that buttons
are actual objects that can be seen, can display text and have spatial
meaning, we are supposed to be getting closer.