Add button clicking functionality
This commit implements clicks on buttons, such that buttons have
actually become clickable objects (in the sense that a click generates a
response). For the current Strata, that means that the Quit and Resume
buttons can now also be clicked, with the same effect as pressing 'q' or
'r'.

The current implementation is not exactly perfect yet, and there may be
good reasons to revise the current way of button clicking. However, the
end of the chain, where a UIReceiver receives an arbitrary function
call, appears to be sufficiently flexible to endure. Also, the receiver
for each button is specified via Lua, which is quite likely the way it's
supposed to be.

In order for Windows to also receive click effects, the chain of events
is rather curious, namely the Window receives a click, which is sent to
the Button, which then goes to the UIInterface, and which then comes
back to the Window (but now as a UIReceiver), after which the click is
interpreted.

It is becoming increasingly attractive to actually collect all the
low-level input in the Window class, and let the derived classes of the
Window only be UIReceivers - and not actually implement input responses,
such as the MainMenu reacting to 'q' and 'r'. It would be more logical
to let 'q' and 'r' be mapped to 'Quit' and 'Resume' (through Lua), and
then let the Window redirect a 'q' button press to a function call to
the MainMenu UIReceiverm, with argument 'Quit'.