Restructure input and bind keys from Lua
Today's commit is the result of considerable trial-and-error on what is
a convenient method for flexible input handling.

Driven by the intention to define more in Lua and less elsewhere, I
defined a text-to-SDLKey conversion function, and I started defining
keys for windows from Lua.

This brings along several changes. The default close-all key SDLK_ESCAPE
is now set from Lua, and so are the commands 'q' and 'r' for
quitting/resuming through the main menu.

I also removed the activeKeys array from the Window, and Windows can now
no longer hardcode which keys they use. (Except the Monitor which still
hardcodes its own trigger key, and the Console which I broke because its
keys aren't loaded from Lua, but I intend to fix that in future
commits.)

Correspondingly, I extended the Lua window and button definitions. I
also removed the getFunctionArgs() function from the Window, opting to
instead let Lua define the function arguments for specific key presses
and button clicks.

In the future, buttons (and maybe keys) might need runtime-dependent
arguments, but the expectation is that these can then be set when the
button is activated / the Window is created.

All of this leaves Strata is a rather messy state and it will take some
more work to finish the move away from hardcoded input handling.