Split Lua UI definition function; use default font size
In this commit I tie up a single loose end from the previous one (namely
the font size, which wasn't set for Buttons and set one-by-one for
Windows from the UI's Font object).

The new way to do this is to simply use the UI's own default values for
both. Passing different values remains possible, but for a consistent UI
feel it's obviously ideal to have a default font size and only deviate
from that if strictly necessary.

Furthermore, I made a separate Lua file for each Window to be created.
This is, in any case, the idea: that every object gets its own
specification.

I also extended the 'loadobj()' function, which loads and executes Lua
functions from specific files. The function now can be given an
arbitrary number of arguments beyond 'path', which allows any kind of
function call and the passing of (a copy of) local variables. This type
of function signature helps reduce the overall size of functions without
having to globally define specific variables in order for them to be
accessible from different functions.