Set keys for monitor/console from Lua
In this commit I did for the Monitor and Console what I already did for
the MainMenu in the previous commit, namely loading the key bindings
from Lua.

When working on the MainMenu yesterday, I was already thinking about how
to load the keys for the Console. After some consideration, I did decide
to not let the Console set its own keys using the key sets of
/src/interface/keys.hpp, but instead load them all from Lua.

The advantages of doing so seem to outweigh the ugliness of listing
every individual key as a 'function' of the Console. Firstly, the
activeKeys (keys used when the window is open/active) are now really not
necessary anymore (they had already been removed two commits ago in
anticipation of this method of implementing the Console). That also
means that key management stays a matter of the Window, and not of its
derived classes, which is much tidier. Secondly, it persists the idea
that Lua is to be used for defining the effects of keys, and not
hardcode them. If someone with a rare keyboard then feels like it, they
can include all the keys on their keyboard to the key set. (Currently
there are some reasons why this isn't possible, primarily because the
mapping from text to SDLKey is still hardcoded and limited, but I hope
to use SDL 2 in the future, and when I do I might allow any kind of
unicode input.)

But, I digress. The Console can again receive input, as it could a while
ago, but the buttons are now clickable and everything is neatly loaded
from Lua. That means the focus can shift first to improving the Console
a bit more, and then towards implementing terrain generation (given that
the last dozen or so commits had nothing to do with terrains, it may be
worthwhile to remind oneself that the actual purpose is not to build a
full-fledged UI but rather to generate terrains, although on the other
hand I've allowed myself to invest so much effort in it because Chathran
will need UI functionality too, and I decided I might as well just build
it all in Strata and extend it only superficially in Chathran).

So, errr, the Console should first be made to display previous comments,
and perhaps it's a nice idea to create a window with clickable buttons
that is used for the terrain generation. After all, console commands are
nice to have, but it might just be convenient to have an overview window
with terrain generation paramaters, which could be quite a bit more
practical than terminal communication alone.

On another note, logging of errors and feedback should also be done via
the Console rather than via std::cout, so I definitely don't want to
suggest that the Console is obsolete before it was fully functional.
It's just that windows are a bit more efficient in managing complex
states.