Shift mouse coordinates to old style, fix click bug
This commit fixes one of Strata's first real bugs: you could actually
click a button even when it was hidden (which obviously I had not tried
yet, but I noticed it from the code).

But, the original reason is that clicking didn't work for the Buttons
anymore because apparently with SDL 2.0, tiny-game-engine clicks are now
in the square [0,2]x[0,2] rather than [-1,1]x[-1,1]. I actually liked
the old style (because (0,0) was the center of the screen), so I simply
added a shift to get the original behavior back.

As a sidenote, I already did invert the 'y' direction from SDL to get
the negative of the vertical direction at the bottom of the screen
rather than the top. So if anyone wonders why the screen coordinates are
a mess... well, sadly, that's just the way it happened to be.