Add base window for Terrain generator
In this commit I mostly add a Lua file for generating an otherwise
unspecial Window, which can hold the functions of the TerrainGenerator.

I want to try to simply create a ui::Window object and then add
sufficient functionality to it not to need to make another derived class
(and yet another special function in the UI). The aim should be to be
able to make various windows with a single function, not to need to
extend the Strata/Lua interface every time a new Window is required.

So far things look good, but then there is no functionality yet. I need
to actually make the TerrainGenerator a class able to receive UI
commands, and then manage to tell the UI from Lua to reach out to the
TerrainGenerator. That, however, is the topic of future commits and this
commit merely adds a nonfunctional extra window.