Update Camera from inside Strata
The UpdateCamera function is now implemented inside Strata, rather than
borrowed from the tiny-game-engine's tiny::os::Application class.

The main reason was that the camera was currently updated from the Game
class, which is too uncontrollable. That is now adjusted as the Camera
is now updated from the UI. Consequently, with this commit, the UI
ensures the Camera is not updated while the Console is visible (i.e.
available for entering text).

Additional criteria are not yet easily possible, and probably would
involve using Lua to specify which windows (if visible) should block the
camera's movement.