Turn Receiver map into a TypeCluster
The setup and usage of the Receiver got into the spotlight when I tried
to add the Terrain as a UIReceiver, and it appears that the UIReceiver
is actually quite like a typical TypeClusterObject: it appears in
various places, is collected in a map containing all such objects
managed by a single class object, needs to subscribe and unsubscribe
itself to that class on creation/destruction... and so on.

Rather than reinvent the wheel, I thus re-used the TypeClusterObject
structure, already in use for the UISource. It looks quite neat and it
even avoids the current issues where receivers and windows are
considered separately. The main change is that I need to also pass a
unique identifier to every Window that is created, but fortunately Lua
already supplies such a unique identifier that is already used for
exactly that purpose... well, that just shows how justified it is to
make this change.

I am not certain whether the UIListener (used by the InputInterpreter)
should be subjected to the same treatment. It's currently also
subscribed and unsubscribed via the UI, and it also uses unique
identifiers... so there's a lot to be said for it. But I need to look
better at how it's being used to see if it is convenient. The InputSet
structure would have to move into the UIListener, at the very least, if
that is going to work.