Enable passing a fixed Renderable index to Renderer
The order in which Renderable objects are rendered is determined by
their index in the tiny-game-engine's Renderer (c.q. WorldRenderer).
That index is also used for freeing Renderable objects, causing them to
no longer be rendered (and removing allocated objects inside the
Renderer, if warranted).

In Strata, Renderable objects are in principle stored in a map as a pair
(Renderable*, unsigned int) where the second member is the key used. One
original aim in creating Window objects for the UI was that one could
re-order them, bringing them forward. This is useful behavior regardless
for Screen renderables, which have no effective 'depth' and thus obscure
each other (and are depending on the order of their indices within the
Renderer).

To allow reordering, one must be able to control the indices. The 'bump'
functionality of the InputInterpreter, in order to actually have some
effect, requires that we free and then re-add the window that is being
clicked on.

When re-adding, the new index must be bigger than all existing indices,
in order to be in front. This requires a currently unused index. One
could go for either keeping two-way maps, or going linearly through the
entire existing map to emulate a two-way map, but for now I simply used
the built-in exception throwing of tiny::draw::Renderer::addRenderable()
which happily throws an exception on an attempt to create a duplicate
entry. The try-catch block in add(World|Screen)RenderableWithIndex()
then ensures nothing is changed in Strata's RenderManager (other than
optionally an increment on the default index counter).

As for the default counter: a while ago I ensured that the
WorldRenderable counter starts at 100000 and the ScreenRenderable
counter starts at 200000. This leaves plenty of room (namely indices
1-99999) to set specific indices for those objects whose ordering is
important. If adding a Renderable fails, it is not retried, so it is
kind of important to ensure that custom indices are not set anywhere
near the default index counters.

Flexible index assignment is also important in future attempts to have
windows be either always-on-top or to avoid the underlying UI from
interfering (e.g. a default, always visible UI, like a HUD-kind of
structure, or a menu bar), as these elements should probably be always
on top or always behind, but either way it should not jump forward on
clicks).