Avoid adding Renderables with duplicate indices
I just extended the tiny-game-engine slightly and initiated a pull
request for the associated commit 5dd66ef3a359b60fc457c4775e9ce8734fe56c21.

For Strata, this extension enables checking whether an index actually
exists before trying to add a Renderable on it. For now, I've simply
added code that tries to find a free index starting from an unsigned int
that's passed to it by reference. If that index is not free, it
increments it, for a hardcoded 1000 times maximum, and rechecks. The
actual index is also returned, such that the caller can ensure that next
time there will be no unnecessary retries on indices that were already
seen to be occupied in a previous attempt.