Add functional 'bump' by bringing window to front
In this commit, a new type of interaction with UI elements is added
through 'bringToFront'. Clicking on a Window does now not only bring it
forward in the sense of input interpretation through the
InputInterpreter (which defines who gets first look when input comes
in), it also reorders the Windows by bringing the clicked Window all the
way forward in the Renderer, such that the window actually appears 'on
top'.

The functionality may be tested by the sequence "Esc, m" and then
clicking the windows at will (obviously, outside of the buttons).