TypeCluster the UIListener
For various reasons this commit follows a two month hiatus. These
reasons include holidays, influenza, december, music, and online games
(in no particular order).  The plan was already to make the UIListener
into a TypeCluster in a late October commit, but lack of actual
programming has thus delayed it until now.

The implementation is straightforward and consists of simple
adjustments. The Window now must use its UIListener's protected
'inputSet' object, as it no longer has direct access to its InputSet.
This is an improvement, since it encapsulates the changes the Window
(and its derivatives) can make to it a little. I've stopped short of
actually forcing the UIListener's derivatives to make a call to get the
InputSet object, which would merely make the necessary code less pretty.

The InputInterpreter now orders a list of listeners, not input sets.
This actually cleans up the InputInterpreter considerably, since it now
only has to order listeners and ask them if they recognize a certain key
input. That's better than the old method, where it all the time needed
to ask the InputSet who its listener was, and then call functions on
that listener in this indirect way.

Finally, I thus adjusted the UIListener and InputSet of the UI
interface. The InputSet is now a true container class, without pointers
to UIListener objects. At the same time, the UIListener is now taking
care of all interface duties UI listening requires.