Send window clicks to Terrain
The code that made a Terrain Generator window was already present, but
its ability to send commands to the Terrain class was not functional (or
lost in a past commit). It is made functional now.

In addition, I adjusted the logic in the Window such that one can now
link key functions to buttons. This allows generic windows to call
functions on key presses. (In the past, windows could only process key
presses through derived classes. Now they only need to have a button,
which defines a function call to any UIReceiver. Ideally, windows should
also be given functionality to receive key input without also having a
corresponding button, but this isn't very urgent.)