Initiate building of vertex map in Terrain
For terrain generation, the first step is to be compression (resulting
in folding). For that to work, we must convert the layered Terrain into
a cloud of points that interact with each other. For the purpose of
compression, we do not need to consider the surface as such, but we
restrict ourselves to seeing all the points as being bound to their
neighbors, with spring-like dynamics.

The current commit starts building a vertex map. It's not used yet, so
it's merely a setup of how to do things. When we have a vertex map, and
each of these vertices has a complete set of neighbors that it is close
to, we can start implementing equations that can then be used to
actually move the vertices around.