Build Vertex map for Terrain generation
In this commit, the code for building a vertex map (partially present in
the previous commit) is completed and used.

On my system, building the vertex map appears to work, but it takes
about 20 seconds to build. This is for about 140,000 vertices, which end
up with about 7 neighbors each (on average) but for which we need to do
no less than 500 million discards (excluding discards due to being far
away), where a neighbor candidate is eventually rejected because it is
already covered by an existing neighbor. The efficiency could
conceivably go up a bit if we'd use smaller meshes (i.e. more
divisions), since right now we apparently need to do about 3500
neighbor-checking loops per vertex (to end up with 7 neighbors).

In future commits I may reorder the logic a bit in order to speed things
up, and perhaps adjust the parameters. However, as the functionality
does appear to work (just not very efficiently), I am already committing
it.