Reorder vertex map logic
I adjusted the method by which the vertex map is built, by first
discarding candidates if they are already covered by existing neighbors
and only then cleanup the neighbors array and add them.

This does not result in any noticeable speedup, but it does make the
logic a bit more straightforward so it is retained anyways.