Reduce max neighbor distance
In this commit I chose the sole remaining method for optimization:
reducing the max distance of neighbors.

In principle the mesh layers are 2.0 apart, and there are 200 divisions
on a mesh of size 300 for a typical separation of 1.5. The initial max
neighbor distance was 30, which was now brought down to 10 without any
observable change in the number of neighbors added. This parameter is
hardcoded, so it's not entirely elegant, but for the time it speeds up
building the vertex map by about half (and cutting the number of vertex
discards down from 500 million to about 50 million).