Mark base layer vertices
In this commit, I added a flag for VertexModifiers that are at the base
of the Terrain (i.e. the lowest Layer), which is set in buildVertexMap.

The equations to determine terrain deformation will rely on applied
forces. The base vertices are special in two ways: firstly, they sense a
net (tangential, usually horizontal) force due to movement of a
hypothetical underlying layer (whereas normal vertices only sense forces
due to gravity and neighboring vertices). Secondly, they are subjected
to a force of buoyancy due to an underlying liquid mass, which
counteracts gravity acting on the Terrain as a whole.