Implement base layer modification, add vertex weights
This commit implements some mechanisms required for calculating forces
acting on the base layer. The buoyancy and drift forces are unique to
the base layer, and do not affect any vertices not in the base layer.

In addition, vertex now have weights (that are set by
Layer::increaseThickness()). By assigning weight to vertices, rather
than tracking a density and multiplying that with volume associated with
a vertex, we avoid having to calculate volume elements during terrain
generation. Instead, we can continue to consider the terrain as a cloud
of vertices on which we aim to calculate a smooth force gradient.

The weight assignment comes down to multiplying a Layer's initial
thickness with the surface associated with that vertex. After this
initial assignment, the aim is to not change weights anymore (absent
erosion/sedimentation).

The surface associated with a vertex is also used dynamically to ensure
that the buoyant force and drift applied to base layer vertices are
proportional to the surface they represent. As compression takes place,
base layer vertices will be squeezed together and the surface per vertex
will decrease. If the buoyancy force would not decrease proportionally,
compressed parts of the Terrain would be subject to an unrealistic
uplifting force.