Project area for buoyancy on y-axis
A minor adjustment to the previous commit, namely to ensure that the
buoyancy on sloped base terrain does not become too big either. In
principle the buoyancy force acts on a horizontal surface, therefore we
now calculate the projection of the normal at a vertex onto the vertical
axis and incorporate that into the calculation of the force.