Add VertexModifier reference to VertexNeighbor
It is starting to look like doing the actual force equilibrium
calculations will be rather involved. In principle, we must aim to do
the terrain modification in a volume-preserving way, which is not
obvious given it's merely a cloud of points.

In any case, we will probably need to do a kind of iterative
equilibration algorithm on the force differences between vertices and
their neighbors. To make this more efficient, in this commit I added a
direct pointer to the VertexModifier to the VertexNeighbor. This helps
avoid making millions of vmap.at() calls to find the neighbor.

According to C++11 specifications quoted on forums, as well as according
to common sense, not only iterators but also references of map elements
are not compromised by insertion or erasure of (other) map elements.
Therefore, the pointer should remain valid as long as its corresponding
map element lives.

I see no way to use a more elegant method of referencing (e.g. using the
iterator doesn't seem possible because the iterator contains a class
(VertexModifier) that depends on the VertexNeighbor). I suppose this is
the way it ought to be done.