Apply compression on Terrain
In this commit, compression has become functional, and terrain
deformation is observed along the direction of compression.

I hardcoded some cappings on how big/small
VertexModifier::forceMultiplier can be. Initially I was too aggressive,
and it turns out using a lower bound of 0.2 (instead of 0.1 as in this
commit) causes the net force to blow up two-way along the compression
direction, presumably because the net force grows
alternating-exponentially.

I am not certain whether this effect should be expected to recur in the
future when other parameters change. I think it ought to be mostly
standalone, depending only on the number of neighbors a vertex typically
has. If neighbor selection is roughly similar to the current situation,
we should not see it again.

Regardless, this commit proves the correctness of several preceding
commits that did not actually make use of newly implemented code. Focus
should now turn towards volume preservation, which is currently totally
absent. The terrain shrinks, but its altitude does not change -
essentially the whole object is just squeezed together along an axis of
convergence, and all movement remains inside the horizontal plane.