Initial improvements to compression methods
There were a couple of changes made (but not committed) before the
preceding unplanned hiatus. I'm committing these changes 'as is' now.
It's not all finished but it doesn't break anything either. The
restorative force that is added here probably improves the results
somewhat.

That said, running this version of Strata will still generate a terrain
that compresses in a rather messy way (with lots of noisy hills formed
even far from the compression axis) and the terrain falling vertically
at the compression axis. While there are no huge excursions, there is a
clear need for further improvement in order for terrains to be more
realistic.