Fix three bugs in compression code
Today's commit takes a few steps back in complexity, as it temporarily
kicks out a handful of influences to compression and removes all but the
base layer. This simplification exposed a handful of weaknesses and
outright mistakes in the compression logic.

Firstly, the restorative force that was the last functional addition
wasn't reset, and it was adjusted in size in a rather arbitrary way.
Secondly, the area was involved in the dragging force of compression,
but not divided out when calculating movement from the 'force'. We have
to start considering 'force' as 'unit force', not something that
actually has magnitude proportional to mass, in at least most of the
situations. Therefore we no longer include 'area', which means the edge
vertices are now compressed as much as internal vertices. Thirdly, and
finally, the calculation for the vector along the axis of zero
compression had an error. For a compression vector (x,z) I had
calculated the perpendicular vector as (x,-z) rather than (z,-x) and
obviously the second kind of perpendicular vector works better.