Use uniform compression rate away from compression line
So far we have always used compression with a rate that linearly
increases with distance from the 'line of zero compression' defined as
the line through center point of the terrain and along the cross product
of the vertical axis and the axis of compression.

This isn't entirely reasonable, though - realistic processes will be
characterised by a mostly linear movement without considerable
compression, and compression takes place only in an area of shortening
such as a mountain range. We don't actually force the absence of
compression elsewhere, we merely apply a constant drift force (as if the
places float on a liquid layer that moves with constant velocity - not a
very unrealistic description) and let the terrain do what it likes as a
consequence.

Given that all forces other than drift remain disabled, this simply
results in a terrain that is linearly drifting towards a line of
convergence, and getting shorter on a band along this line.

This commit also fixes a bug where maxMeshSize was used where the
terrain size was intended. The terrain size, a variable that didn't
exist previously, measures the size of the terrain at creation (before
compression). The max mesh size is a technical parameter that sets the
maximal dimension of Bundle and Strip objects that make up a Layer. This
is usually related to the Terrain size, but substantially lower.