Add buoyancy/gravity and fix altitude bug
This commit reenables buoyancy and gravity. Again, for consistent force
calculation, we divide out the area to get a more dimensionless force
(that acts upon a region, basically an infinitesimal (but layer-thick)
piece of the layer rather than upon a certain amount of mass).

In addition, I fixed an issue that caused compressive force to increase
with altitude (it ignored the vertical component of position in
determining the distance to the compression axis).